# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CrewTally API
servers:
  - url: /api
    description: Replit API proxy prefix
...
paths:
...
  /v1/auth/apple:
    post:
      operationId: signInWithApple
      tags: [auth]
      summary: Verify Apple sign-in and create a session
      description: Verifies the Apple identity token and issues an opaque CrewTally session token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleSignInRequest"
      responses:
        "200":
          description: Session created
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleSignInResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /v1/auth/dev:
    post:
      operationId: signInDev
      tags: [auth]
      summary: Sign in as a configured development owner
      description: Available only when APP_ENV is development and DEV_SIGNIN_CODE is configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevSignInInput"
      responses:
        "200":
          description: Session created
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleSignInResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /v1/auth/signout:
    post:
      operationId: signOut
      tags: [auth]
      summary: Revoke the caller's session
      description: Revokes only the session associated with the presented bearer token.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /v1/invitations:
    parameters: *workspaceParameters
    get:
      operationId: listInvitations
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Pending and recent invitations; never raw credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationList" }
    post:
      operationId: createInvitation
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationInput" }
      responses:
        "200":
          description: Credentials returned once; replay returns null credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationCreated" }
  /v1/invitations/{id}:
    parameters:
      - { $ref: "#/components/parameters/WorkspaceHeader" }
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: revokeInvitation
      security: [{ bearerAuth: [] }]
      requestBody: *operationBody
      responses:
        "200": *okResponse
  /v1/invite/peek:
    post:
      operationId: peekInvitation
      requestBody: &tokenBody
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationTokenInput" }
      responses:
        "200":
          description: Workspace name and role only, or unavailable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPreview" }
  /v1/invite/accept:
    post:
      operationId: acceptInvitation
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationAcceptInput" }
      responses:
        "200": &joinedResponse
          description: Membership created or reactivated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationAccepted" }
        "410":
          description: Invitation not available
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/invite/accept-code:
    post:
      operationId: acceptInvitationCode
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationCodeInput" }
      responses:
        "200": *joinedResponse
        "400":
          description: Wrong, expired or rate-exhausted code; one opaque answer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/invite/decline:
    post:
      operationId: declineInvitation
      requestBody: *tokenBody
      responses:
        "200": *okResponse
components:
  parameters:
    WorkspaceHeader:
      name: X-Workspace-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    InvitationInput:
      type: object
      required: [operation_id, role, email]
      properties:
        operation_id: { type: string, format: uuid }
        role: { type: string, enum: [PARTNER, ADMIN, LEAD, WORKER] }
        email: { type: string, format: email, maxLength: 254 }
        name: { type: string, maxLength: 60 }
        financial_access: { type: [boolean, "null"] }
        worker_id: { type: [string, "null"], format: uuid }
      additionalProperties: false
    Invitation:
      type: object
      required: [id, invitee_name, email, role, status, expires_at]
      properties:
        id: { type: string, format: uuid }
        invitee_name: { type: [string, "null"] }
        email: { type: string, format: email }
        role: { type: string }
        status: { type: string, enum: [PENDING, ACCEPTED, REVOKED, DECLINED, EXPIRED] }
        expires_at: { type: string, format: date-time }
    InvitationList:
      type: object
      required: [invitations]
      properties:
        invitations:
          type: array
          items: { $ref: "#/components/schemas/Invitation" }
    InvitationCreated:
      type: object
      required: [id, expires_at, token, code, link, already_created]
      properties:
        id: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
        token: { type: [string, "null"] }
        code: { type: [string, "null"] }
        link: { type: [string, "null"] }
        already_created: { type: boolean }
    InvitationTokenInput:
      type: object
      required: [token]
      properties:
        token: { type: string, pattern: "^[A-Za-z0-9_-]{43}$" }
      additionalProperties: false
    InvitationAcceptInput:
      type: object
      required: [token, operation_id]
      properties:
        token: { type: string, pattern: "^[A-Za-z0-9_-]{43}$" }
        operation_id: { type: string, format: uuid }
      additionalProperties: false
    InvitationCodeInput:
      type: object
      required: [email, code, operation_id]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        code: { type: string, pattern: "^[0-9]{6}$" }
        operation_id: { type: string, format: uuid }
      additionalProperties: false
    InvitationPreview:
      type: object
      required: [available]
      properties:
        available: { type: boolean }
        workspace_name: { type: string }
        role: { type: string }
    InvitationAccepted:
      type: object
      required: [workspace_id, role, already_member]
      properties:
        workspace_id: { type: string, format: uuid }
        role: { type: string }
        already_member: { type: boolean }
...
    AppleSignInRequest:
      type: object
      properties:
        identityToken:
          type: string
          minLength: 1
          maxLength: 16384
        authorizationCode:
          type: string
          minLength: 1
          maxLength: 4096
        rawNonce:
          type: string
          minLength: 43
          maxLength: 43
          pattern: "^[A-Za-z0-9_-]{43}$"
      required: [identityToken, authorizationCode, rawNonce]
      additionalProperties: false
    DevSignInInput:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 4096
        label:
          type: string
          enum: [owner-a, owner-b, member-c, member-d]
      required: [code, label]
      additionalProperties: false
    AppleSignInResponse:
      type: object
      properties:
        sessionToken:
          type: string
          minLength: 43
          maxLength: 43
        user:
          $ref: "#/components/schemas/User"
      required: [sessionToken, user]
      additionalProperties: false
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
      required: [id]
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            correlationId:
              type: string
              format: uuid
          required: [code, message, correlationId]
          additionalProperties: false
      required: [error]
      additionalProperties: false
  responses:
    BadRequest:
      description: Malformed or invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Invalid identity or session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Development sign-in is not enabled in this environment
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too many authentication requests from this IP
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Safe server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"