# lib/api-zod/src/generated/types/invitationCodeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CrewTally API
 * OpenAPI spec version: 0.1.0
 */

export interface InvitationCodeInput {
  /** @maxLength 254 */
  email: string;
  /** @pattern ^[0-9]{6}$ */
  code: string;
  operation_id: string;
}