# lib/api-zod/src/generated/types/invitationList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CrewTally API
 * OpenAPI spec version: 0.1.0
 */
import type { Invitation } from './invitation';

export interface InvitationList {
  invitations: Invitation[];
}